backend/requirements.txt
django
djangorestframework
requests
django-cors-headers
django-redis
redis-server